Conflicts:
	backend/src/IdleLogManager.cc
	frontend/common/src/System.cc

+++ backend/include/IStatistics.hh
@@ -21,6 +21,7 @@
 #define ISTATISTICS_HH
 
 #include <time.h>
+#include <stdint.h>
 
 #include "ICore.hh"
 
@@ -53,7 +54,7 @@ namespace workrave {
       };
 
     typedef int BreakStats[STATS_BREAKVALUE_SIZEOF];
-    typedef int MiscStats[STATS_VALUE_SIZEOF];
+    typedef uint64_t MiscStats[STATS_VALUE_SIZEOF];
 
     struct DailyStats
     {
